requirements.txt
Pillow
# tkinter
PyQt5
opencv-python
pyautogui
keyboard